implementation/app/requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-Bcrypt==1.0.*
pytest==9.0.*
# PostgreSQL adapter for Python
psycopg2-binary==2.9.*
Flask-JWT-Extended==4.7.*
requests==2.32.*
pytest-cov==7.0.*